+++ src/main/java/de/fanta/cubeside/mixin/MixinGameModeSwitch.java
@@ -1,5 +1,6 @@
 package de.fanta.cubeside.mixin;
 
+import com.mojang.blaze3d.platform.InputConstants;
 import de.fanta.cubeside.config.Configs;
 import net.minecraft.client.KeyboardHandler;
 import net.minecraft.client.Minecraft;
@@ -16,7 +17,7 @@ public class MixinGameModeSwitch {
     @Inject(at = @At("HEAD"), method = "handleDebugKeys(Lnet/minecraft/client/input/KeyEvent;)Z", cancellable = true)
     public void behaviour(KeyEvent input, CallbackInfoReturnable<Boolean> cir) {
         if (Configs.Generic.GamemodeSwitcher.getBooleanValue()) {
-            if (input.input() == 293) {
+            if (input.input() == InputConstants.KEY_F4) {
                 Minecraft.getInstance().setScreenAndShow(new GameModeSwitcherScreen());
                 cir.setReturnValue(true);
             }